pull out a couple common types
