+++ lib/wificaptive/src/wifi-types.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <Arduino.h>
+
+struct WifiCredentials
+{
+    String ssid;
+    String pswd;
+};
+
+struct Network
+{
+    String ssid;
+    int32_t rssi;
+    bool open;
+    bool saved;
+};
